Add solution to Money Tree problem
Add the solution to the Money Tree problem, which resolves the tree by
determining the last line first and building the tree in reverse, until
all rows / lines of the tree have been recorded. As each line is
recorded the money symbol decreases in size by a half, down to a minimum
of one `2n +1`, whilst simultaneously the edges increase by one each row
heading up `n`. Therefore a loop the counts up (or down) increases the
padding from zero on either side, whilst decreasing the repetition of
the money symbol `$` in the middle.