Add initial solution to Caesar Cipher problem
Add an initial solution to Caesar Cipher problem that relies on a
`shiftAlphabet` function to build a cipher from a maybe given array of
alphabet letters and shifts it by 3, when no number provided. The uses a
`decrypt` method to find and replace letters within encrypted message
with those from the cipher, with respect to the provided alphabet of
letters.