Add `timer` object to re-implement `timers.js`
Add the `timer` object to reimplement the behaviour within `timers.js`,
but keep it separate within its own `timer.js` script in order to keep
the original as is. This reimplements the Stop Watch behaviours for each
button, but within the `var timer` variable (for better or worse) as
requested by the extension activity.