Update Event handling to Parent not Child in ToDo
Update the event handling within the To Do application by applying the
Event Listener to the parent DOM element, rather than to each child
element, but instead requires additional _type_ or _class_ checking to
ensure that the ToDo List is not being marked as "completed" and instead
the child elements ToDo Item are.

By applying the event listener once to the parent element means that it
no longer requires event listeners to be attached to each child element
within the parent element for the same effect as before.