Add Week 04 Day 2 Ruby Warmup exercise
Add the week four day two part one Ruby warmup exercise to have a
conversation with Daniel (formerly "Bob") where the method response with
a sentence of text depending on questions or remarks asked act them.
This is includes an initial solution before attempting the extension
with l33t speak.