Add solution to "99 Bottles of Beer" sinatra app
Add a solution to the "99 Bottles of Beer" Sinatra application, which
relies on named parameters within the route to determine the route taken
and contents of the page, providing a page with the number of bottles on
the wall with a link to take the next one down (decrease it by one).

Also, adds a route for `0` bottles to explain no bottles on the wall and
link to start again.