Update styles for Movie template on mobile
Update the styles to display the Movie poster behind (and slightly over)
the initial details when viewed on mobile, so that the poster takes up
the full width of the screen, fades into the text and does not squash
the text elements beside and instead shifts them to underneath it.
Otherwise on larger screens increase the poster size as the screen size
increases and revert the details back to the old/current styles.