Refactor common operators into define_method
Refactor the common operators into a `define_method` block that sends
the operator to the first argument with the other value and substitutes
in the operator, since the body of work is the same for each operator
expect for the operator being used, hence the refactoring.

This will also allow for easier access to add more operators should they
be needed.

== References:
- https://www.mutuallyhuman.com/blog/2011/01/25/class-coercion-in-ruby
- https://stackoverflow.com/questions/2799571/in-ruby-how-does-coerce-actually-work
- https://ruby-doc.org/core-2.2.0/Module.html#method-i-define_method