Refactor Movie searching into Service object
Refactor the call to the external Movie Database service via a Service
object that contains the External API request and have it return a
response that be can manipulate to fill out the template. Otherwise,
return an empty like response to fill in the template on first view,
after search and if search failed.