Add an idea of Relations for Repositories
Add an idea of building a Relation object for chaining queries on a
given scope of a table without directly using `sql`. A Relation for a
Table can be built, have its parameters of `where`, `limit` and `order`
modified and return the SQL results for that query, when turned into an
array / collection object in Ruby.