Update WDI City Pix project with part 2 solution
Update the WDI City Pix project with the solution to part 2 and update
the background of the page using the provided CSS classes on the
document body tag / element. This change needed more information from
the initial collection, which was transformed into an object to use its
keys as both values and keys to solve the problem.